fairy/text2speech/requirements.txt
elevenlabs
fastapi
gunicorn
python-dotenv
python-multipart
requests
uvicorn